Refactor code
Organise based on business structure

Refactor API url handling
