--- frontend/src/service/api.ts
@@ -1,5 +1,0 @@
-import axios from "axios";
-
-export const api = axios.create({
-  baseURL: "http://localhost:3000",
-});
