+++ frontend/src/service/questionApiClient.ts
@@ -1,0 +1,5 @@
+import axios from "axios";
+
+export const questionApiClient = axios.create({
+  baseURL: import.meta.env.VITE_API_URL,
+});
