+++ frontend/src/service/questionsService.ts
@@ -1,11 +1,11 @@
 import { Question, QuestionWithOptional } from "../types/question";
-import { api } from "./api";
+import { questionApiClient } from "./questionApiClient";
 
 const QUESTION_API_URL = "/questions";
 
 export const getQuestions = async () => {
   try {
-    const response = await api.get(QUESTION_API_URL);
+    const response = await questionApiClient.get(QUESTION_API_URL);
     return response.data;
   } catch (error) {
     console.error("Error fetching questions: ", error);
@@ -15,7 +15,7 @@ export const getQuestions = async () => {
 
 export const addQuestion = async (question: QuestionWithOptional) => {
   try {
-    const response = await api.post(QUESTION_API_URL, question);
+    const response = await questionApiClient.post(QUESTION_API_URL, question);
     console.log("Question added: ", response.data);
     return response.data;
   } catch (error) {
@@ -26,7 +26,7 @@ export const addQuestion = async (question: QuestionWithOptional) => {
 
 export const editQuestion = async (question: Question) => {
   try {
-    const response = await api.put(
+    const response = await questionApiClient.put(
       `${QUESTION_API_URL}/${question.id}`,
       question
     );
@@ -40,7 +40,9 @@ export const editQuestion = async (question: Question) => {
 
 export const deleteQuestion = async (questionId: number) => {
   try {
-    const response = await api.delete(`${QUESTION_API_URL}/${questionId}`);
+    const response = await questionApiClient.delete(
+      `${QUESTION_API_URL}/${questionId}`
+    );
     console.log("Question deleted: ", response.data);
     return response.data;
   } catch (error) {
