+++ frontend/src/types/question.ts
@@ -9,5 +9,6 @@ export interface Question {
 // This type is used in the form to allow id to be optional when creating a new question
 export type QuestionWithOptional = Omit<Question, "id" | "complexity"> & {
   id?: number;
-  complexity?: "" | "Easy" | "Medium" | "Hard";
+  complexity?: "Easy" | "Medium" | "Hard";
+  friendlyID?: number;
 };
