+++ services/question/src/index.ts
@@ -14,9 +14,11 @@ const app = express()
 const port = process.env.PORT || 3000
 
 app.use(express.json())
-app.use(cors({
-  origin: process.env.CORS_ORIGINS ? JSON.parse(process.env.CORS_ORIGINS) : '*'
-}))
+app.use(
+  cors({
+    origin: process.env.CORS_ORIGINS ? JSON.parse(process.env.CORS_ORIGINS) : '*',
+  })
+)
 
 app.use('/questions', router)
